video_summarizer/requirements.txt
torch
torchvision
torchaudio
openai-whisper
numpy
opencv-python
faiss-cpu
transformers
librosa
tqdm
moviepy
langchain
safetensors